Cover declared env lowering
