+++ tests/build_graph.rs
@@ -4,6 +4,9 @@ use zen::build_graph::{
 use zen::lexer;
 use zen::parser;
 
+#[path = "build_graph/host_effects.rs"]
+mod host_effects;
+
 fn parse_program(src: &str) -> zen::ast::Program {
     let tokens = lexer::tokenize(src, 0).expect("lex build script");
     parser::parse(tokens, 0).expect("parse build script")
@@ -427,70 +430,5 @@ build = (b: Builder) Result<BuildConfig, BuildError> {
     assert_eq!(
         err.to_string(),
         "build target `app` cannot depend on itself"
-    );
-}
-
-#[test]
-fn build_program_lowering_rejects_undeclared_env_reads() {
-    let program = parse_program(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    std_path = b.os.env("ZEN_STD")
-    b.add(Executable { name: "myapp", main: "main.zen", out_dir: "build/" })
-    .Ok(b.config())
-}
-"#,
-    );
-
-    let err = BuildGraph::from_build_program(&program)
-        .expect_err("undeclared build.zen env read should fail");
-
-    assert_eq!(
-        err.to_string(),
-        "undeclared host effect: read env `ZEN_STD`"
-    );
-}
-
-#[test]
-fn build_program_lowering_accepts_declared_file_reads() {
-    let program = parse_program(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    manifest = b.os.read_file("build.targets") ?
-        | .Ok(contents) { contents }
-        | .Err { "default" }
-    b.add(Executable { name: "myapp", main: "main.zen", out_dir: "build/" })
-    .Ok(b.config())
-}
-"#,
-    );
-
-    let graph = BuildGraph::from_build_program(&program).expect("lower build graph");
-    let json = graph.canonical_json().expect("build graph json");
-
-    assert!(
-        json.contains(r#""kind":"read_file","value":"build.targets""#),
-        "expected read-file host effect in graph json, json={json}"
-    );
-}
-
-#[test]
-fn build_program_lowering_rejects_undeclared_file_reads() {
-    let program = parse_program(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    manifest = b.os.read_file("build.targets")
-    b.add(Executable { name: "myapp", main: "main.zen", out_dir: "build/" })
-    .Ok(b.config())
-}
-"#,
-    );
-
-    let err = BuildGraph::from_build_program(&program)
-        .expect_err("undeclared build.zen file read should fail");
-
-    assert_eq!(
-        err.to_string(),
-        "undeclared host effect: read file `build.targets`"
     );
 }
