+++ tests/docs_truth/phase_audit.rs
@@ -35,6 +35,7 @@ fn phase_plan_records_recovered_progress_and_next_slice() {
         "build_program_lowering_rejects_unknown_target_dependencies",
         "build_program_lowering_rejects_unsupported_package_targets",
         "build_program_lowering_rejects_unsupported_link_targets",
+        "build_program_lowering_accepts_declared_env_reads",
         "build_program_lowering_rejects_undeclared_env_reads",
         "emit_json_build_graph_outputs_project_build_graph",
         "emit_json_build_graph_outputs_library_target",
@@ -215,6 +216,7 @@ fn completion_audit_records_recovery_objective_evidence_and_gaps() {
         "build_program_lowering_rejects_unknown_target_dependencies",
         "build_program_lowering_rejects_unsupported_package_targets",
         "build_program_lowering_rejects_unsupported_link_targets",
+        "build_program_lowering_accepts_declared_env_reads",
         "emit_json_build_graph_outputs_library_target",
         "emit_json_build_graph_outputs_target_dependencies_and_features",
         "emit_json_build_graph_rejects_cyclic_target_dependencies",
